refactor(desktop): extract the pure sort model out of smart-sort-bar
