+++ apps/desktop/src/renderer/src/lib/__tests__/sort-model.test.ts
@@ -4,7 +4,7 @@ import {
   toggleColumnSort,
   type SortChip,
   type SortColumn
-} from '@/components/smart-sort-bar'
+} from '@/lib/sort-model'
 
 type ChipInit = {
   id?: string
